Introduces support for AWT's component
Fixes #76
