+++ jsvg/src/main/java/com/github/weisj/jsvg/SVGDocument.java
@@ -27,6 +27,7 @@ import java.awt.geom.Path2D;
 
 import javax.swing.*;
 
+import com.github.weisj.jsvg.renderer.awt.AwtComponentPlatformSupport;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -36,7 +37,6 @@ import com.github.weisj.jsvg.geometry.size.FloatSize;
 import com.github.weisj.jsvg.geometry.size.MeasureContext;
 import com.github.weisj.jsvg.nodes.SVG;
 import com.github.weisj.jsvg.renderer.*;
-import com.github.weisj.jsvg.renderer.awt.JComponentPlatformSupport;
 import com.github.weisj.jsvg.renderer.awt.NullPlatformSupport;
 import com.github.weisj.jsvg.renderer.awt.PlatformSupport;
 
@@ -74,18 +74,27 @@ public final class SVGDocument {
     }
 
     public void render(@Nullable JComponent component, @NotNull Graphics2D graphics2D, @Nullable ViewBox bounds) {
+        render(component, graphics2D, bounds);
+    }
+
+    public void render(@Nullable Component component, @NotNull Graphics2D graphics2D, @Nullable ViewBox bounds) {
         PlatformSupport platformSupport = component != null
-                ? new JComponentPlatformSupport(component)
+                ? new AwtComponentPlatformSupport(component)
                 : new NullPlatformSupport();
+        renderWithPlatform(platformSupport, graphics2D, bounds);
+    }
+
+    private float computePlatformFontSize(@NotNull PlatformSupport platformSupport, @NotNull Output output) {
+        return output.contextFontSize().orElseGet(platformSupport::fontSize);
+    }
+
+    public void renderWithPlatform(@NotNull PlatformSupport platformSupport, @NotNull Graphics2D graphics2D,
+                                   @Nullable ViewBox bounds) {
         Graphics2D g = (Graphics2D) graphics2D.create();
         setupSVGRenderingHints(g);
         Output output = new Graphics2DOutput(g);
         renderWithPlatform(platformSupport, output, bounds);
         output.dispose();
-    }
-
-    private float computePlatformFontSize(@NotNull PlatformSupport platformSupport, @NotNull Output output) {
-        return output.contextFontSize().orElseGet(platformSupport::fontSize);
     }
 
     public void renderWithPlatform(@NotNull PlatformSupport platformSupport, @NotNull Output output,
